requirements.txt
openai>=1.6.0
httpx>=0.27.0
typer>=0.9.0
python-dotenv>=1.0.0
tweepy>=4.14.0
